rename just scorer (instead of base_scorer)
