+++ weave/flow/eval.py
@@ -11,7 +11,7 @@ from typing_extensions import Self
 
 import weave
 from weave.flow import util
-from weave.flow.base_scorer import (
+from weave.flow.scorer import (
     Scorer,
     _has_oldstyle_scorers,
     _validate_scorer_signature,
