+++ weave/scorers/__init__.py
@@ -1,4 +1,4 @@
-from weave.flow.base_scorer import (
+from weave.flow.scorer import (
     Scorer,
     _has_oldstyle_scorers,
     _validate_scorer_signature,
